+++ Library/OcAfterBootCompatLib/OcAfterBootCompatLib.c
@@ -19,14 +19,14 @@
 #include <Library/OcDebugLogLib.h>
 #include <Library/DevicePathLib.h>
 #include <Library/MemoryAllocationLib.h>
-#include <Library/OcAppleBootCompatLib.h>
+#include <Library/OcAfterBootCompatLib.h>
 #include <Library/OcDevicePathLib.h>
 #include <Library/OcMiscLib.h>
 #include <Library/PrintLib.h>
 #include <Library/UefiRuntimeServicesTableLib.h>
 #include <Library/UefiBootServicesTableLib.h>
 
-#include <Protocol/OcAppleBootCompat.h>
+#include <Protocol/OcAfterBootCompat.h>
 
 #include "BootCompatInternal.h"
 
@@ -35,15 +35,15 @@
   with legacy AptioMemoryFix protocol, allowing us to avoid
   conflicts between the two.
 **/
-STATIC OC_APPLE_BOOT_COMPAT_PROTOCOL mOcAppleBootCompatProtocol = {
-  OC_APPLE_BOOT_COMPAT_PROTOCOL_REVISION
+STATIC OC_AFTER_BOOT_COMPAT_PROTOCOL mOcAfterBootCompatProtocol = {
+  OC_AFTER_BOOT_COMPAT_PROTOCOL_REVISION
 };
 
 /**
   Apple Boot Compatibility context. This context is used throughout
   the library. Must be accessed through GetBootCompatContext ().
 **/
-STATIC BOOT_COMPAT_CONTEXT  mOcAppleBootCompatContext;
+STATIC BOOT_COMPAT_CONTEXT  mOcAfterBootCompatContext;
 
 STATIC
 EFI_STATUS
@@ -56,7 +56,7 @@ InstallAbcProtocol (
   EFI_HANDLE  Handle;
 
   Status = gBS->LocateProtocol (
-    &gOcAppleBootCompatProtocolGuid,
+    &gOcAfterBootCompatProtocolGuid,
     NULL,
     &Interface
     );
@@ -73,8 +73,8 @@ InstallAbcProtocol (
   Handle = NULL;
   Status = gBS->InstallMultipleProtocolInterfaces (
     &Handle,
-    &gOcAppleBootCompatProtocolGuid,
-    &mOcAppleBootCompatProtocol,
+    &gOcAfterBootCompatProtocolGuid,
+    &mOcAfterBootCompatProtocol,
     NULL
     );
 
@@ -91,7 +91,7 @@ GetBootCompatContext (
   VOID
   )
 {
-  return &mOcAppleBootCompatContext;
+  return &mOcAfterBootCompatContext;
 }
 
 EFI_STATUS
@@ -124,9 +124,9 @@ OcAbcInitialize (
 
   DEBUG ((
     DEBUG_INFO,
-    "OCABC: FEXITBS %d PRCSM %d CSLIDE %d PRSRV %d RBMAP %d VMAP %d APPLOS %d RTPERMS %d\n",
+    "OCABC: FEXITBS %d PRMRG %d CSLIDE %d PRSRV %d RBMAP %d VMAP %d APPLOS %d RTPERMS %d\n",
     Settings->ForceExitBootServices,
-    Settings->ProtectCsmRegion,
+    Settings->ProtectMemoryRegions,
     Settings->ProvideCustomSlide,
     Settings->ProtectUefiServices,
     Settings->RebuildAppleMemoryMap,
@@ -136,7 +136,7 @@ OcAbcInitialize (
     ));
 
   DEBUG_CODE_BEGIN ();
-  TotalMemory = CountFreePages (&LowMemory);
+  TotalMemory = OcCountFreePages (&LowMemory);
   DEBUG ((
     DEBUG_INFO,
     "OCABC: Firmware has %Lu free pages (%Lu in lower 4 GB)\n",