+++ Changelog.md
@@ -20,10 +20,12 @@ OpenCore Changelog
 - Update builtin firmware
 - Fixed `PowerTimeoutKernelPanic` on 10.15.4
 - Fixed 4K section alignment in `OpenRuntime` to fix Linux booting on SKL
-- Introduced `SyncRuntimePermissions` to fix Linux booting on CFL+
+- Introduced `SyncRuntimePermissions` to fix multiple memory permission flaws
 - Introduced `RebuildAppleMemoryMap` to fix macOS booting on Dell 5490
 - Removed `ShrinkMemoryMap` in favour of more advanced `RebuildAppleMemoryMap`
 - Marked `EnableWriteUnprotector` as deprecated on modern systems
+- Introduced `ProtectMemoryRegions` to fix memory region handling
+- Removed `ProtectCsmRegion` in favour of `ProtectMemoryRegions`
 
 #### v0.5.6
 - Various improvements to builtin text renderer